Experiment: Subset fonts (#32)
* Add subfont for the index page

* Cache subfonts

* Correctly include assets

* Wording

* Use uglier URLs to allow subfont to run

* Revert 'about' changes

* Fix the /posts/ assetgraph error

Turns out it could not resolve href="/posts/" for the link rel.
I am not sure why. My theory is that ag has different heuristics
for directories (directory/index.html) for link[rel] as opposed to
[href] relations. It was thus looking for _site/posts/ and could
not find it! I fixed it by making the path absolute, which seem
like a better approach anyway.

* Expand comment :)

* Add webmentions back in